+++ src/ListenerProvider/SubscribableTrait.php
@@ -21,13 +21,14 @@ trait SubscribableTrait
      * @template T of object
      * @param class-string<T>|null $type
      * @param ?callable(T|Emitted<T>): void $listener
+     * @param string|array<string>|null $action
      * @return Subscription<T>
      */
     public function on(
         ?string $type = null,
         ?callable $listener = null,
         ?string $context = null,
-        ?string $action = null,
+        string|array|null $action = null,
         Priority $priority = Priority::Medium
     ): Subscription {
         $subscription = $this->createSubscription(
@@ -49,13 +50,14 @@ trait SubscribableTrait
      * @template T of object
      * @param class-string<T>|null $type
      * @param ?callable(T|Emitted<T>): void $listener
+     * @param string|array<string>|null $action
      * @return Subscription<T>
      */
     public function once(
         ?string $type = null,
         ?callable $listener = null,
         ?string $context = null,
-        ?string $action = null,
+        string|array|null $action = null,
         Priority $priority = Priority::Medium
     ): Subscription {
         $subscription = $this->createSubscription(
@@ -78,13 +80,14 @@ trait SubscribableTrait
      * @template T of object
      * @param class-string<T>|null $type
      * @param ?callable(T|Emitted<T>): void $listener
+     * @param string|array<string>|null $action
      * @return Subscription<T>
      */
     public function createSubscription(
         ?string $type = null,
         ?callable $listener = null,
         ?string $context = null,
-        ?string $action = null,
+        string|array|null $action = null,
         Priority $priority = Priority::Medium,
         bool $singleUse = false
     ): Subscription {